// src/models/RedditModels/AllSubredditModels.ts
export interface Child {
    kind: string;
    data: Data2;
}

export interface Data {
    after: string;
    dist: number;
    modhash: string;
    geo_filter?: any;
    children: Child[];
    before?: any;
}

export interface SubredditResponse {
    kind: string;
    data: Data;
}