// src/models/RedditModels/AllSubredditPostModels.ts
export interface Child {
    kind: string;
    data: Data2;
}

export interface Data {
    after?: any;
    dist?: number;
    modhash: string;
    geo_filter: string;
    children: Child[];
    before?: any;
}

export interface SubredditPostResponse {
    kind: string;
    data: Data;
}